requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
duckduckgo_search==6.1.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
groq==0.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
logger==1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.26.4
ollama==0.2.0
openai==1.30.1
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pgvector==0.2.5
phidata==2.4.10
pillow==10.3.0
protobuf==4.25.3
psycopg==3.1.19
psycopg-binary==3.1.19
pyarrow==16.1.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
pypdf==4.2.0
pyreqwest_impersonate==0.4.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
streamlit==1.35.0
tenacity==8.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
